Correct smoothing range parameter (#1150)
The smoothing range parameter used for the eigenvalue estimation was defined as integer while it should be a double.

+++ include/core/parameters.h
@@ -1137,7 +1137,7 @@ namespace Parameters
     bool mg_smoother_eig_estimation;
 
     /// MG smoothing range to set range between eigenvalues
-    int eig_estimation_smoothing_range;
+    double eig_estimation_smoothing_range;
 
     /// MG number of cg iterations to find eigenvalue
     int eig_estimation_cg_n_iterations;
